equiz jari's push

+++ app/src/main/java/com/teamx/equiz/data/models/getcart/GetCartData.kt
@@ -5,7 +5,7 @@ import androidx.annotation.Keep
 @Keep
 
 data class GetCartData(
-    val cartPrice: Int,
+    val cartPrice: Double,
     val `data`: List<Data>,
     val discount: Double,
     val discountCoupon: DiscountCoupon
